Moved around some of the code so that the functions can be reused by other cloud components
